feat(android): persist check items and seed step2 demo data
Wire check_items through Room and sale persistence, switch reports to date-range source data, trigger sync after successful sale, and add deterministic seed/report tests to lock the Step 2 local-first data path.

+++ android/app/src/main/java/com/vitbon/kkm/features/sales/domain/SalesModels.kt
@@ -11,7 +11,7 @@ data class CartItem(
     val discount: Money = Money.ZERO,
     val vatRate: VatRate
 ) {
-    val total: Money get() = Money(price.kopecks * quantity) - discount
+    val total: Money get() = Money((price.kopecks * quantity).toLong()) - discount
 }
 
 data class Cart(
